chore: organise tests by type and add pytest markers
