+++ core/tests/selenium/__init__.py
@@ -1,0 +1,1 @@
+"""Selenium-Tests."""
