+++ core/tests/unit/test_anlage3_parser.py
@@ -1,21 +1,15 @@
 from pathlib import Path
 from tempfile import NamedTemporaryFile
 
-from django.core.files.uploadedfile import SimpleUploadedFile
-from docx import Document
-
-from .base import NoesisTestCase
 import pytest
-from pathlib import Path
-from tempfile import NamedTemporaryFile
-from docx import Document
 from django.core.files.uploadedfile import SimpleUploadedFile
+from docx import Document
 
-from .base import NoesisTestCase
-from ..models import BVProject, BVProjectFile
-from ..anlage3_parser import parse_anlage3
+from ..base import NoesisTestCase
+from ...models import BVProject, BVProjectFile
+from ...anlage3_parser import parse_anlage3
 
-pytestmark = pytest.mark.usefixtures("seed_db")
+pytestmark = [pytest.mark.unit, pytest.mark.usefixtures("seed_db")]
 
 
 class Anlage3ParserTests(NoesisTestCase):
