+++ pytest.ini
@@ -1,9 +1,11 @@
 [pytest]
 DJANGO_SETTINGS_MODULE = noesis.settings
 pythonpath = .
-addopts = --reuse-db -ra -n auto
+addopts = --reuse-db -ra -n auto -m "unit or integration and not slow"
 testpaths = core/tests
 markers =
     selenium: runs selenium browser tests
     e2e: marks end-to-end tests
     slow: marks tests as slow
+    unit: marks unit tests
+    integration: marks integration tests
